feat(settings): add mobile Email & Password tab, generalize account-access
Split account-access's dialog-card chrome from its reusable menu/email/
password content so a new sheet-only tab can render the same flow the
aside's Edit button opens on tablet/desktop. Moves modals/account-access
to settings/account-access to reflect it's settings-owned.
