+++ src/components/settings/account-access/email-section.vue
