+++ src/components/settings/account-access/password-section.vue
