+++ src/locales/en-us.json
@@ -151,6 +151,7 @@
   "settings.tab.app": "App Settings",
   "settings.tab.review-preferences": "Study Settings",
   "settings.tab.danger-zone": "Danger Zone",
+  "settings.tab.account-access": "Email & Password",
   "settings.header.index.title": "Settings",
   "settings.index.account-heading": "Account",
   "settings.index.app-heading": "Settings",
